Reorganize data type codec tests
