--- src/test/java/io/reactiverse/pgclient/DataTypeTestBase.java
@@ -1,29 +1,0 @@
-package io.reactiverse.pgclient;
-
-import io.vertx.core.Vertx;
-import io.vertx.ext.unit.TestContext;
-import io.vertx.ext.unit.junit.VertxUnitRunner;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.runner.RunWith;
-
-/**
- * @author <a href="mailto:emad.albloushi@gmail.com">Emad Alblueshi</a>
- */
-@RunWith(VertxUnitRunner.class)
-public abstract class DataTypeTestBase extends PgTestBase {
-
-  Vertx vertx;
-
-  protected abstract PgConnectOptions options();
-
-  @Before
-  public void setup() {
-    vertx = Vertx.vertx();
-  }
-
-  @After
-  public void teardown(TestContext ctx) {
-    vertx.close(ctx.asyncAssertSuccess());
-  }
-}
