+++ src/test/java/io/reactiverse/pgclient/codec/PreparedStatementParamCoercionTest.java
@@ -1,5 +1,9 @@
-package io.reactiverse.pgclient;
+package io.reactiverse.pgclient.codec;
 
+import io.reactiverse.pgclient.PgClient;
+import io.reactiverse.pgclient.PgConnectOptions;
+import io.reactiverse.pgclient.PgConnection;
+import io.reactiverse.pgclient.Tuple;
 import io.reactiverse.pgclient.data.Numeric;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
