+++ src/test/java/io/reactiverse/pgclient/codec/extended/BinaryDataTypesTest.java
@@ -1,0 +1,57 @@
+package io.reactiverse.pgclient.codec.extended;
+
+import io.reactiverse.pgclient.PgClient;
+import io.reactiverse.pgclient.Row;
+import io.reactiverse.pgclient.Tuple;
+import io.reactiverse.pgclient.codec.ColumnChecker;
+import io.reactiverse.pgclient.codec.ExtendedQueryDataTypeCodecTestBase;
+import io.vertx.core.buffer.Buffer;
+import io.vertx.ext.unit.Async;
+import io.vertx.ext.unit.TestContext;
+import org.junit.Test;
+
+import java.util.Random;
+
+public class BinaryDataTypesTest extends ExtendedQueryDataTypeCodecTestBase {
+  @Test
+  public void testBytea(TestContext ctx) {
+    Random r = new Random();
+    int len = 2048;
+    byte[] bytes = new byte[len];
+    r.nextBytes(bytes);
+    Async async = ctx.async();
+    PgClient.connect(vertx, options, ctx.asyncAssertSuccess(conn -> {
+      conn.prepare("SELECT $1::BYTEA \"Bytea\"",
+        ctx.asyncAssertSuccess(p -> {
+          p.execute(Tuple.of(Buffer.buffer(bytes)), ctx.asyncAssertSuccess(result -> {
+            ColumnChecker.checkColumn(0, "Bytea")
+              .returns(Tuple::getValue, Row::getValue, Buffer.buffer(bytes))
+              .returns(Tuple::getBuffer, Row::getBuffer, Buffer.buffer(bytes))
+              .forRow(result.iterator().next());
+            async.complete();
+          }));
+        }));
+    }));
+  }
+
+  @Test
+  public void testBufferArray(TestContext ctx) {
+    Random r = new Random();
+    int len = 2048;
+    byte[] bytes = new byte[len];
+    r.nextBytes(bytes);
+    Async async = ctx.async();
+    PgClient.connect(vertx, options, ctx.asyncAssertSuccess(conn -> {
+      conn.prepare("SELECT ARRAY[$1::BYTEA] \"Bytea\"",
+        ctx.asyncAssertSuccess(p -> {
+          p.execute(Tuple.of(Buffer.buffer(bytes)), ctx.asyncAssertSuccess(result -> {
+            ColumnChecker.checkColumn(0, "Bytea")
+              .returns(Tuple::getValue, Row::getValue, new Buffer[]{Buffer.buffer(bytes)})
+              .returns(Tuple::getBufferArray, Row::getBufferArray, new Buffer[]{Buffer.buffer(bytes)})
+              .forRow(result.iterator().next());
+            async.complete();
+          }));
+        }));
+    }));
+  }
+}
